+++ index.js
@@ -42,6 +42,7 @@ const safeObject = require('./util/safe_object');
  * @property {false} [caniuse] Disable CanIUse embeds.
  * @property {false} [youtube] Disable YouTube embeds.
  * @property {false} [wistia] Disable Wistia embeds.
+ * @property {false} [twitter] Disable Twitter embeds.
  * @property {false} [underline] Disable underline syntax.
  * @property {false|import('./modifiers/fence_label').FenceLabelOptions} [fence_label] Disable fence labels, or set options for the feature.
  * @property {false|import('./modifiers/fence_secondary_label').FenceSecondaryLabelOptions} [fence_secondary_label] Disable fence secondary labels, or set options for the feature.
@@ -137,6 +138,10 @@ module.exports = (md, options) => {
 
     if (optsObj.wistia !== false) {
         md.use(require('./rules/embeds/wistia'), safeObject(optsObj.wistia));
+    }
+
+    if (optsObj.twitter !== false) {
+        md.use(require('./rules/embeds/twitter'), safeObject(optsObj.twitter));
     }
 
     // Register modifiers